Abstract out utility functions from index file
